Fix terminal_cron.sh not being added to the crontab
After many, many hours of debugging and isolating the problem down to
just three lines of code: the curl download, the un-tarring and
finally, adding the line to the crontab, the only solution I've come
up with is... to run the line twice. For some reason it doesn't work
on the first try, but it works on the second. Why? Who knows.