Move path up sooner in bash_profile
If the path is too far down in the list, version checks (like the
setting of the color flags in 'ls') will be done against different
executables. Therefore, we should move it up after the options.

Also move functions up before aliases, because we might want to alias
a function.