Fix bug with install script
The `confirm` lib function expects three parameters:

 * The text to display for the confirmation
 * (Optional) Variable to check to bypass confirmation
 * (Optional) Text to display if confirmed

If the second parameter is an unquoted string, the variable is expanded before the shell executes the command. Ordinarily this wouldn't be a problem, unless a third (non-empty) parameter is used. When that happens:

```
confirm 'Would you like to confirm?' $VAR 'Doing action'
```

becomes:

```
confirm 'Would you like to confirm'  'Doing action'
```

This means that `confirm` will see a non-empty second parameter, which causes the function to short circuit. Since the value is very unlikely to be Y/y, the return value will be 0 and any blocks attached to an `if` will be bypassed.