Fix bug in piping through sudo tee
Because we have aliased `sudo` to `validate_sudo || init_sudo; sudo `
piping anything into the sudo command actually pipes it into
`validate_sudo` and not the command `sudo`. As such it was hanging
when it reached this point in the code.

Ideally we should figure out a way to pass piped commands through sudo
as this bug will probably hit us again. Or maybe we should throw an
error when there is standard input into `validate_sudo`. But, it needs
to work *now* and not later.