Add a sudo invalidation to our validate_sudo function
Although it sounds out, it's necessary to invalidate the sudo
*timestamp* before we try to check to see if our password is working.
The reason for this is if the sudo timestamp is still active, then
the subsequent call to `sudo -v` will succeed, even if the password we
passed in on stdin is wrong.

If we do not invalidate the timestamp, then `validate_sudo` succeeds
during `init_sudo`. Because we want to force the user to enter a
password when `init_sudo` is called and not the first time we hit
`sudo` in the script, we need this to fail.